requirements.txt
anthropic>=0.97.0
chromadb>=1.5.8
fastapi>=0.136.1
httpx>=0.28.1
openai>=2.33.0
pydantic>=2.13.3
pydantic-settings>=2.14.0
python-socketio>=5.16.1
uvicorn>=0.46.0
